move the new component to adminconsole
